requirements.txt
requests>=2.28.0
pandas>=1.5.0
numpy>=1.24.0
solana>=0.30.0
solders>=0.18.0
python-telegram-bot>=20.0
web3>=6.0.0
cryptography>=3.4.0
python-dateutil>=2.8.0
pytz>=2022.7
urllib3>=1.26.0
aiohttp>=3.8.0
httpx>=0.24.0
colorama>=0.4.6
python-dotenv>=1.0.0